pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "p2p-energy-backend"
version = "0.1.0"
description = "Django 5 + DRF + SimpleJWT backend for P2P Energy Swap"
requires-python = ">=3.10"
dependencies = [
  "Django>=5.0,<6.0",
  "djangorestframework>=3.15",
  "djangorestframework-simplejwt>=5.3",
  "django-cors-headers>=4.4",
  "psycopg[binary]>=3.1",
  "python-dotenv>=1.0",
  "Pillow>=10.0"
]

[tool.setuptools]
packages = {find = {}}
